Fixed various bugs with new cigars
cigar-fields.json AND mappedRange.js: Changed Strength Profile null value to an empty string. This is due to the bug detailed in #68

submit.js: When a cigar doesn't have a match in the original data, it now returns an empty string instead of undefined, which causes errors with the diff library. This should only apply in the case of new cigars added to the local data.

crudForm.js: when finding unique values in subarrays, empty subarrays are filtered first